customer micro service done

+++ CustomerMS/src/main/java/com/cg/dao/CustomerDAO.java
@@ -9,6 +9,6 @@ import com.cg.entity.Customer;
 @Repository("CustomerDAO")
 public interface CustomerDAO extends JpaRepository<Customer, Integer>{
 	@Query("select c from Customer c where c.customerId = ?1")
-	public Customer findCustomerById(long l);
+	public Customer findCustomerById(int l);
 
 }
